Changement structure + intégration barre de recherche
