+++ client/src/components/HeaderNavMobile.jsx
@@ -1,16 +1,46 @@
-import { Link } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
+import PropTypes from "prop-types";
 import "../styles/HeaderNavMobile.scss";
 import { useClickAway } from "@uidotdev/usehooks";
-import { useRef, useState } from "react";
+import { useRef, useState, useEffect } from "react";
 import { Squash as Hamburger } from "hamburger-react";
 import { AnimatePresence, motion } from "framer-motion";
+
 import routes from "../routes";
 import logo from "../assets/images/GlobPrewiew.png";
 
-function HeaderNavMobile() {
+function HeaderNavMobile({ selectionManager, destinations }) {
+  const location = useLocation();
+
   const [isOpen, setOpen] = useState(false);
+
   const ref = useRef(null);
   useClickAway(ref, () => setOpen(false));
+
+  const [selection, setSelection] = useState("");
+
+  const navSelectedCountry = destinations.map((data) => (
+    <option value={data.Name} key={data.Name}>
+      {data.Name}
+    </option>
+  ));
+
+  useEffect(() => {
+    // early return
+    if (!selection) {
+      return;
+    }
+
+    const indexOfSelection = destinations.findIndex(
+      (d) => d.Name === selection
+    );
+
+    if (indexOfSelection !== undefined) {
+      const destinationToReturn = destinations[indexOfSelection];
+
+      selectionManager.manageCountrySelection(destinationToReturn);
+    }
+  }, [selection]);
 
   return (
     <div className="HearderContainer">
@@ -66,13 +96,32 @@ function HeaderNavMobile() {
       <div className="HearderTitle">
         <p>Bienvenue chez</p>
         <h1>GLOBE GUIDE</h1>
-        <input
-          className="headerInput"
-          type="text"
-          placeholder="Search destination"
-        />
+        {location.pathname === "/" && (
+          <div className="DinputBtn">
+            <form>
+              <select
+                name="countrie"
+                id=""
+                className="DheaderInput"
+                onChange={(e) => setSelection(e.target.value)}
+              >
+                <option value="">choose your country</option>
+                {navSelectedCountry}
+              </select>
+            </form>
+          </div>
+        )}
       </div>
     </div>
   );
 }
+
+HeaderNavMobile.propTypes = {
+  destinations: PropTypes.arrayOf.isRequired,
+  selectionManager: PropTypes.shape({
+    selectedCountry: PropTypes.string,
+    manageCountrySelection: PropTypes.func.isRequired,
+  }).isRequired,
+};
+
 export default HeaderNavMobile;
