+++ client/src/main.jsx
@@ -4,27 +4,44 @@ import App from "./App";
 import Home from "./pages/Home";
 import AboutList from "./pages/AboutList";
 import Contact from "./pages/Contact";
+import CountrySelectionManager from "./components/CountrySelectionManager";
+import destinations from "./data/destinations";
 
-const root = ReactDOM.createRoot(document.getElementById("root"));
+function Main() {
+  const selectionManager = new CountrySelectionManager();
 
-const router = createBrowserRouter([
-  {
-    element: <App />,
-    children: [
-      {
-        path: "/",
-        element: <Home />,
-      },
-      {
-        path: "/about",
-        element: <AboutList />,
-      },
-      {
-        path: "/contact",
-        element: <Contact />,
-      },
-    ],
-  },
-]);
+  const router = createBrowserRouter([
+    {
+      element: (
+        <App selectionManager={selectionManager} destinations={destinations} />
+      ),
+      children: [
+        {
+          path: "/",
+          element: (
+            <Home
+              selectionManager={selectionManager}
+              destinations={destinations}
+            />
+          ),
+        },
+        {
+          path: "/about",
+          element: <AboutList />,
+        },
+        {
+          path: "/contact",
+          element: <Contact />,
+        },
+      ],
+    },
+  ]);
 
-root.render(<RouterProvider router={router} />);
+  return (
+    <div>
+      <RouterProvider router={router} />
+    </div>
+  );
+}
+
+ReactDOM.createRoot(document.getElementById("root")).render(<Main />);
